// libs/juridico/src/lib/models/SeccionInformativa/infCasInformativaModel.ts
export interface InfCasInformativaModel {
   refTabla: string;
   numFormul: string;
   mtoCas801: number;
   mtoCas803: number;
   mtoCas210: number;
   mtoCas216: string;
   mtoCas217: number;
   mtoCas221: number;
   mtoCas222: string;
   mtoCas804: number;
   mtoCas802: number;
   mtoCas199: number;
   mtoCas687: string;
   mtoCas208: string;
   mtoCas207: string;
   mtoCas815: string;
   mtoCas814: string;
   mtoCas209: number;
   mtoCas211: string;
   mtoCas215: number;
   mtoCas289: number;
   mtoCas290: number;
   mtoCas291: number;
   mtoCas225: string;
   mtoCas226: string;
   mtoCas817: string;
   mtoCas818: string;
   mtoCas829: number;
   mtoCas293: string;
   mtoCas250: string;
   mtoCas251: string;
   mtoCas252: string;
   mtoCas253: string;
   mtoCas254: string;
   mtoCas255: string;
   mtoCas256: string;
   mtoCas257: string;
   mtoCas258: string;
   mtoCas259: string;
   mtoCas260: string;
   mtoCas261: string;
   mtoCas262: string;
   mtoCas263: string;
   mtoCas821: number;
   mtoCas213: number;
   mtoCas240: number;
   mtoCas241: number;
   mtoCas242: number;
   mtoCas243: number;
   mtoCas819: number;
   mtoCas227: number;
   mtoCas502: number;
   mtoCas782: number;
   mtoCas824: number;
}